// packages/browser/ac-browser/src/lib/ac-inputs/enums/ac-enum-input-event.enum.ts
export enum AcEnumInputEvent{
  Blur = 'blur',
  Change = 'change',
  Click = 'click',
  Destroy = 'destroy',
  DoneTyping = 'donetyping',
  DoubleClick = 'dblclick',
  Focus = 'focus',
  Init = 'init',
  Input = 'input',
  KeyDown = 'keydown',
  KeyPress = 'keypress',
  KeyUp = 'keyup',
  MouseDown = 'mousedown',
  MouseEnter = 'mouseenter',
  MouseLeave = 'mouseleave',
  MouseOut = 'mouseout',
  MouseOver = 'mouseover',
  MouseUp = 'mouseup',
  ValueChange = 'valuechange'
}